refactor: Player와 Dealer의 중복 메서드 추상 클래스로 분리
