+++ src/main/java/blackjack/domain/Dealer.java
@@ -1,36 +1,13 @@
 package blackjack.domain;
 
-import java.util.List;
 
-public class Dealer {
-    private final Cards drawnCards;
-
+public class Dealer extends Participant{
     public Dealer() {
-        drawnCards = new Cards();
-    }
-
-    public void receiveOneCard(Card card) {
-        drawnCards.addCard(card);
-    }
-
-    public Cards getDrawnCards() {
-        return drawnCards;
-    }
-
-    public List<String> getCardNames() {
-        return drawnCards.getCardNames();
+        super();
     }
 
     public int getAdditionalDrawnCardCount() {
         return drawnCards.getCardNames().size() - 2;
-    }
-
-    public boolean isBust() {
-        return drawnCards.sumScore() > 21;
-    }
-
-    public int calculateTotalScore() {
-        return drawnCards.sumScore();
     }
 
     public boolean isDealerDone() {
