+++ src/main/java/blackjack/domain/Player.java
@@ -1,38 +1,14 @@
 package blackjack.domain;
 
-import java.util.List;
-
-public class Player {
+public class Player extends Participant {
     private final String name;
 
-    private final Cards drawnCards;
-
     public Player(String name) {
+        super();
         this.name = name;
-        this.drawnCards = new Cards();
     }
 
     public String getName() {
         return name;
-    }
-
-    public void receiveOneCard(Card card) {
-        drawnCards.addCard(card);
-    }
-
-    public Cards getDrawnCards() {
-        return drawnCards;
-    }
-
-    public List<String> getCardNames() {
-        return drawnCards.getCardNames();
-    }
-
-    public boolean isBust() {
-        return drawnCards.sumScore() > 21;
-    }
-
-    public int calculateTotalScore() {
-        return drawnCards.sumScore();
     }
 }
